filled in most command function files.
